git-svn-id: https://svn.apache.org/repos/asf/sis/branches/JDK6@1551742 13f79535-47bb-0310-9956-ffa450edef68

+++ core/sis-utility/src/main/java/org/apache/sis/util/ArgumentChecks.java
@@ -182,7 +182,7 @@ public final class ArgumentChecks extends Static {
         if (value != null) {
             final Class<?> valueClass = value.getClass();
             if (!expectedType.isAssignableFrom(valueClass)) {
-                final int key;
+                final short key;
                 final Object[] args;
                 if (name != null) {
                     key = Errors.Keys.IllegalArgumentClass_3;
